Rejigged the validators module to always internally use to_python.
formencode's validation mechanism performs a greater number of checks when
using to_python compared to from_python - this is especially the case when
you try to use schema validation and want default values to be filled in.

While it would be nice to use one validator and use the natural input / output
mechanism - in this case, it's easier to just use the same validator in the
same direction, and then to slap value-to-text conversion wrappers around the
validator when pushing the values out to SOAP.

Also got more involved with using meaningful, easier to grok identifiers for
validators to make it easier to inspect what the values are expected to be.